Create Readme.cmd
The NHS is facing challenges in managing its growing patient demand and optimising its 
resource utilisation across a wide network of services. The increasing population in England 
requires careful budget allocation to expand NHS infrastructure while ensuring the existing 
resources are used efficiently. A particular focus is on understanding the causes of missed GP 
appointments, as this results in avoidable costs.  
1. Provide actionable insights to inform NHS decisions.  
2. Help NHS scale infrastructure for growing demand.  
3. Guide NHS budget allocation through utilisation trends.  
4. Analyse Twitter data to inform public sentiment and budget decisions.