Add files via upload
Customer Demographics of 2 Market,  Tableau visualisation.

Note: In the Tableau dashboard for “Customer Demographics of 2Market,” SP was initially labelled as São Paulo. This was an error — SP refers to Spain in this context.
The technical report has been updated to reflect this correction. Apologies for any confusion caused.